requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
sqlalchemy==2.0.45
asyncpg==0.31.0
redis==5.2.1
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.12.0
mcp==1.24.0
sse-starlette==3.0.3